+++ core/structs.go
@@ -28,20 +28,39 @@ func (Data *TlData) RefreshFeeds() error {
 }
 
 type TlConfig struct {
-	Subscribed_data         string
-	Date_format             string
-	Log_file                string
-	Highlights              string
-	Cli_limit               int
-	Mode                    string
+	// Global config:
+	Subscribed_data string
+	Date_format     string
+	Log_file        string
+	Highlights      string
+	Mode            string
+	// Tinylog file edition:
+	Allow_edit        bool
+	Tinylog_path      string
+	Post_edit_script  string
+	Post_edit_refresh bool
+	// CLI options:
+	Cli_limit int
+	// TUI options:
 	Tui_status_emoji        bool
-	Allow_edit              bool
-	Tinylog_path            string
-	Post_edit_script        string
-	Post_edit_refresh       bool
 	Tui_max_entries         int
 	Tui_copy_stub_clipboard bool
 	Tui_show_stub           bool
+	// TUI theme:
+	Tui_color_background                 string
+	Tui_color_links                      string
+	Tui_color_text                       string
+	Tui_color_focus_box                  string
+	Tui_color_author_name                string
+	Tui_color_elapsed_time               string
+	Tui_color_box                        string
+	Tui_color_selected_background        string
+	Tui_color_selected_foreground        string
+	Tui_color_highlight                  string
+	Tui_color_quote                      string
+	Tui_color_button_text                string
+	Tui_color_button_selected_background string
+	Tui_color_button_background          string
 }
 
 type TlFeed struct {